move FFI functions to separate file
